Fixes and test update
